Move ImageCacheKey into the image module
That way it'll be eaiser to deal with the code with image-decoders disabled
