+++ internal/backends/gl/images.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 use std::rc::Rc;
 
 use i_slint_core::api::euclid;
-use i_slint_core::graphics::{cache::ImageCacheKey, IntSize, SharedImageBuffer};
+use i_slint_core::graphics::{ImageCacheKey, IntSize, SharedImageBuffer};
 use i_slint_core::{items::ImageRendering, ImageInner};
 
 use super::glrenderer::{CanvasRc, GLItemRenderer};
